Move files around per nits
