+++ test/PartyBid/Bid.test.js
@@ -8,7 +8,7 @@ const { eth, contribute, bidThroughParty } = require('../helpers/utils');
 const { placeBid } = require('../helpers/externalTransactions');
 const { deployTestContractSetup } = require('../helpers/deploy');
 const { MARKETS, MARKET_NAMES } = require('../helpers/constants');
-const { testCases } = require('../partyBidTestCases.json');
+const { testCases } = require('./partyBidTestCases.json');
 
 describe('Bid', async () => {
   MARKETS.map((marketName) => {
