+++ test/PartyBid/Contribute.test.js
@@ -11,7 +11,7 @@ const {
 } = require('../helpers/utils');
 const { deployTestContractSetup } = require('../helpers/deploy');
 const { MARKETS } = require('../helpers/constants');
-const { testCases } = require('../partyBidTestCases.json');
+const { testCases } = require('./partyBidTestCases.json');
 
 describe('Contribute', async () => {
   MARKETS.map((marketName) => {
